add cast

+++ cublibx/srcs/private/_cublx_pset.c
@@ -6,7 +6,7 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/22 20:52:59 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/08/24 18:34:44 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/08/25 18:54:31 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@ int	_cublx_pset(t_cublx *self, int x, int y, int color)
 
 	view = self->main_view[self->view_switch];
 	dst = view.addr + (y * view.line_length + x * (view.bits_per_pixel / 8));
-	*(unsigned int *)dst = color;
+	*(unsigned int *)dst = (unsigned int)color;
 	return (0);
 }
